requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
httpx
pydantic
passlib[bcrypt]
python-jose
python-multipart
pytest